clean up entry points and iron up output the code structure
* add equalibrim energy of birch-murnaghan fit to delta factor workflow.
* unified all parameters name to V0 B0 B1.
* remove unused WF entry points from setup.py
* also remove useless example code and example entry points
* refactoring helper functions in to one place
