+++ setup.json
@@ -21,27 +21,18 @@
     "version": "0.1.0b0",
     "entry_points": {
         "aiida.data": [
-            "sssp_workflow = aiida_sssp_workflow.data:DiffParameters"
         ],
         "aiida.calculations": [
-            "sssp_workflow = aiida_sssp_workflow.calculations.calculations:DiffCalculation",
             "sssp_workflow.birch_murnaghan_fit = aiida_sssp_workflow.calculations.birch_murnaghan_fit:birch_murnaghan_fit",
             "sssp_workflow.calculate_delta = aiida_sssp_workflow.calculations.calculate_delta:calculate_delta",
             "sssp_workflow.calculate_delta_volume = aiida_sssp_workflow.calculations.calculate_delta_volume:calculate_delta_volume"
         ],
         "aiida.parsers": [
-            "sssp_workflow = aiida_sssp_workflow.parsers:DiffParser"
         ],
         "aiida.cmdline.data": [
-            "sssp_workflow = aiida_sssp_workflow.cli.cli:data_cli"
         ],
         "aiida.workflows": [
-            "sssp_workflow.eos = aiida_sssp_workflow.workflows.eos:EquationOfStateWorkChain",
             "sssp_workflow.delta_factor = aiida_sssp_workflow.workflows.delta_factor:DeltaFactorWorkChain",
-            "sssp_workflow.cohesive_energy = aiida_sssp_workflow.workflows.cohesive_energy:CohesiveEnergyWorkChain",
-            "sssp_workflow.pressure_evaluation = aiida_sssp_workflow.workflows.pressure:PressureEvaluationWorkChain",
-            "sssp_workflow.phonon_frequencies = aiida_sssp_workflow.workflows.phonon_frequencies:PhononFrequenciesWorkChain",
-            "sssp_workflow.bands = aiida_sssp_workflow.workflows.bands:BandsWorkChain",
             "sssp_workflow.convergence.cohesive_energy = aiida_sssp_workflow.workflows.convergence.cohesive_energy:ConvergenceCohesiveEnergyWorkChain",
             "sssp_workflow.convergence.phonon_frequencies = aiida_sssp_workflow.workflows.convergence.phonon_frequencies:ConvergencePhononFrequenciesWorkChain",
             "sssp_workflow.convergence.pressure = aiida_sssp_workflow.workflows.convergence.pressure:ConvergencePressureWorkChain",
